Segment Tree

+++ Leecode Daily Problem/1894. Find the Student that Will Replace the Chalk.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int chalkReplacer(vector<int>& chalk, int k) {
 
-        long long sum = accumulate(chalk.begin(), chalk.end(), 0LL);
+        int sum = accumulate(chalk.begin(), chalk.end(), 0LL);
         k %= sum;
         for(int i = 0; i < chalk.size(); i++){
             if(chalk[i] > k)   return i;
